Make timeout longer and quit on mistake
I took the exam, and did not pass it.
The reason was I was too eager to go through, even though I needed to yield.
The project probably had an influence.

To make me focus on choosing the correct action, instead of
reacting unnecessarily fast, I use a greater reaction time (20s), and
I impose a bigger penalty on mistakes, namely quitting the application.

In addition, the scoring system is changed, showing you the total of correct moves,
instead of some arbitrary reaction-time based score.